self-audit: 8 quality fixes across PTY + bridge + a2a-client work
Industrial-textbook pass over the four PTY/bridge/client commits I
landed this session. Two are bugs; the rest are correctness +
clarity drift caught re-reading the code with hostile eyes.

Bugs (behavior-affecting):

  pty_attach_ability writer (T2) — protocol violation:
  Stdin frame missing `data` field used to emit a fake `stdout`
  frame with an `_error` sidecar. An MCP client treating that as
  terminal output would print `<empty>` to the user. Now emits a
  `warn` frame type per §D5; documented in the preamble.

  pty_attach_ability writer (T2) — silent write failures:
  `let _ = tokio::task::spawn_blocking(...).await;` ate both the
  JoinError and the inner io::Result. The loop continued eating
  frames against a dead handle. Now matches on the inner Result
  and breaks the loop on any non-Ok-Ok-(); drop the to_client
  clone, let T3's exit-frame + §I2 quorum close cleanly.

Lying docs:

  pty_attach_ability preamble: documented exit frame as fired
  "when the child terminates" but T1's PTY-cant-lend-reader path
  also emitted exit{status:null} with no child involved. Now
  documents both paths and the `warn` frame type.

  pty/mod.rs::close: claimed "kill() sends SIGKILL on unix" —
  verified against portable-pty 0.8 src/unix.rs that it actually
  sends SIGHUP. Doc now matches reality.

  a2a_client_ability::DISPATCHER_HANDLE: doc referenced a function
  `set_dispatcher_for_test` that doesn't exist. Removed; clarified
  that the populated-path test belongs at the daemon-bin
  integration level (would race the static OnceLock from a unit
  test).

Surface drift:

  pty_lifecycle_ability error messages: dropped six redundant
  `pty_session_create:` / `pty_session_close:` prefixes — same
  SR-6+9 lesson the Go half learned earlier (the dispatcher wraps
  with the ability name; double-prefix is noise + a rename hazard).

  a2a_client_ability::send_task_handler: dropped a double-prefix
  `format!("remote dispatch failed: {e}")` — `e` already names
  the failure (e.g. "NoopGateway: ..."), the prefix doubled up in
  the rendered string.

  pty/mod.rs::PtyCloseOutcome.exit_status: was `Option<i32>`. The
  `as i32` cast on portable-pty's `u32 exit_code()` was a footgun
  for extended-status fields (high bit). Now `Option<u32>`,
  preserving the source type.

  pty_attach_ability T3 exit-watcher: emitted `Some(-1)` as a
  sentinel for "wait error → treat as exit." `-1` collides with
  nothing legal but conveys nothing either; replaced with `null`
  on the wire (consistent with T1's PTY-cant-lend-reader path).

Cosmetic:

  pty_attach_ability::attach_handler: deleted a four-line
  drop-then-re-clone dance with an apologetic comment explaining
  the bug to myself. Clean shape: clone twice from
  xport_from_handler_tx, move it once into the writer. Same
  number of senders.

  mcp_bridge_ability::success_response: replaced
  `unwrap_or_else(|_| "<unencodable>".into())` with `.expect(...)`.
  serde_json::to_string on a Value cannot fail (no NaN, no cycles);
  the fallback was dead code.

  a2a_client_ability::send_task_handler: extracted three near-
  identical required-string-field validation blocks into one
  helper `required_nonempty_string`. 18 lines → 9, adding a
  fourth field would now be one line.

  shell_command() test fixture: returned `&'static str`, forcing
  three call sites to write `Some(shell_command().to_string())`.
  Now returns String; call sites are `Some(shell_command())`.

Tests: 51 across pty_attach + pty_lifecycle + pty service +
mcp_bridge + a2a_client + a2a_bridge + agents conformance, ran
5x parallel = 5/5 deterministic green. The full lib sweep has 2
failures in `support::shellguard::destructive` (pre-existing
Windows-path basename bug in user's untracked WIP, not introduced
by this commit).

+++ src/runtime/execution/pty/mod.rs
@@ -151,9 +151,10 @@ pub struct PtyCloseOutcome {
     /// (idempotent close — the surface is "remove if present").
     pub ack: bool,
     /// Exit status reported by the child, when waitable. None when
-    /// the row was unknown OR when the child was force-killed
-    /// before wait could grab a status.
-    pub exit_status: Option<i32>,
+    /// the row was unknown OR when the child was killed before wait
+    /// could grab a status. u32 preserves portable-pty's own type;
+    /// POSIX exit codes are 0-255 in practice.
+    pub exit_status: Option<u32>,
 }
 
 /// Process-wide PTY session registry. Cloneable handle (the inner
@@ -253,13 +254,16 @@ impl PtyService {
     ///
     /// Sequence:
     ///   1. Remove the session row (no further attach can claim it).
-    ///   2. Take the child out of its slot, kill it. We don't wait
-    ///      forever for a graceful exit because the close path is
-    ///      synchronous from the wire's perspective; portable-pty's
-    ///      kill() sends SIGKILL on unix, which is decisive.
+    ///   2. Take the child out of its slot, kill it. portable-pty's
+    ///      Child::kill() sends SIGHUP on unix (verified against
+    ///      portable-pty 0.8 src/unix.rs) — most shells handle that
+    ///      as "session disconnected" and clean up gracefully. We
+    ///      don't escalate to SIGKILL because the close path is
+    ///      synchronous from the wire's perspective; a stuck child
+    ///      can be reaped by the OS later.
     ///   3. Try-wait once for the exit status; if it's not yet
-    ///      reaped (rare — kill is synchronous on unix) return
-    ///      `exit_status: None`. The OS reaper still claims it.
+    ///      reaped return `exit_status: None`. The OS reaper still
+    ///      claims it.
     pub fn close(&self, id: &PtySessionId) -> PtyCloseOutcome {
         let session = {
             let mut g = self.inner.lock().expect("pty service lock");
@@ -280,9 +284,13 @@ impl PtyService {
                 exit_status: None,
             };
         };
-        // SIGKILL on unix; ignored if the child already exited.
+        // SIGHUP on unix; ignored if the child already exited.
         let _ = child.kill();
-        let exit_status = child.try_wait().ok().flatten().map(|s| s.exit_code() as i32);
+        // exit_code() returns u32; preserve as u32 (POSIX exit codes
+        // are 0-255). The previous `as i32` cast was a footgun for
+        // the rare cases when extended status fields encode signal
+        // info in high bits.
+        let exit_status = child.try_wait().ok().flatten().map(|s| s.exit_code());
         PtyCloseOutcome {
             ack: true,
             exit_status,
